apply the loadbalancer concurrency fix to the kubernetes client
