+++ dataclients/kubernetes/kube_test.go
@@ -375,7 +375,8 @@ func TestIngressData(t *testing.T) {
 		},
 		[]*ingressItem{testIngress("foo", "baz", "bar", "", "", "", backendPort{8080}, 1.0)},
 		map[string]string{
-			"kube_foo__baz______": "http://1.2.3.4:8080",
+			"kube_foo__baz______":           "http://1.2.3.4:8080",
+			"kube_foo__baz________lb_group": "",
 		},
 	}, {
 		"service backend from ingress and service, path rule",
@@ -403,7 +404,8 @@ func TestIngressData(t *testing.T) {
 			),
 		)},
 		map[string]string{
-			"kube_foo__baz__www_example_org_____bar": "http://1.2.3.4:8181",
+			"kube_foo__baz__www_example_org_____bar":           "http://1.2.3.4:8181",
+			"kube_foo__baz__www_example_org_____bar__lb_group": "",
 		},
 	}, {
 		"service backend from ingress and service, default and path rule",
@@ -432,8 +434,10 @@ func TestIngressData(t *testing.T) {
 			),
 		)},
 		map[string]string{
-			"kube_foo__qux______":                    "http://1.2.3.4:8080",
-			"kube_foo__qux__www_example_org_____baz": "http://5.6.7.8:8181",
+			"kube_foo__qux______":                              "http://1.2.3.4:8080",
+			"kube_foo__qux________lb_group":                    "",
+			"kube_foo__qux__www_example_org_____baz":           "http://5.6.7.8:8181",
+			"kube_foo__qux__www_example_org_____baz__lb_group": "",
 		},
 	}, {
 		"service backend from ingress and service, with port name",
@@ -461,7 +465,8 @@ func TestIngressData(t *testing.T) {
 			),
 		)},
 		map[string]string{
-			"kube_foo__qux__www_example_org_____bar": "http://1.2.3.4:8181",
+			"kube_foo__qux__www_example_org_____bar":           "http://1.2.3.4:8181",
+			"kube_foo__qux__www_example_org_____bar__lb_group": "",
 		},
 	}, {
 		"ignore ingress entries with missing metadata",
@@ -505,7 +510,8 @@ func TestIngressData(t *testing.T) {
 			},
 		},
 		map[string]string{
-			"kube_foo__qux__www_example_org_____bar": "http://1.2.3.4:8181",
+			"kube_foo__qux__www_example_org_____bar":           "http://1.2.3.4:8181",
+			"kube_foo__qux__www_example_org_____bar__lb_group": "",
 		},
 	}} {
 		t.Run(ti.msg, func(t *testing.T) {
@@ -737,17 +743,26 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__default_only______":                           "http://1.2.3.4:8080",
-			"kube_namespace2__path_rule_only__www_example_org_____service3": "http://9.0.1.2:7272",
-			"kube_namespace1__mega______":                                   "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__foo_example_org____":                          "",
-			"kube_namespace1__mega__bar_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__bar_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__bar_example_org____":                          "",
-			"kube_namespace1__ratelimit______":                              "http://1.2.3.4:8080",
-			"kube_namespace1__ratelimitAndBreaker______":                    "http://1.2.3.4:8080",
+			"kube_namespace1__default_only______":                                     "http://1.2.3.4:8080",
+			"kube_namespace1__default_only________lb_group":                           "",
+			"kube_namespace2__path_rule_only__www_example_org_____service3":           "http://9.0.1.2:7272",
+			"kube_namespace2__path_rule_only__www_example_org_____service3__lb_group": "",
+			"kube_namespace1__mega______":                                             "http://1.2.3.4:8080",
+			"kube_namespace1__mega________lb_group":                                   "",
+			"kube_namespace1__mega__foo_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__foo_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__foo_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__foo_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__foo_example_org____":                                    "",
+			"kube_namespace1__mega__bar_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__bar_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__bar_example_org____":                                    "",
+			"kube_namespace1__ratelimit______":                                        "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimit________lb_group":                              "",
+			"kube_namespace1__ratelimitAndBreaker______":                              "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimitAndBreaker________lb_group":                    "",
 		})
 	})
 
@@ -774,7 +789,8 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__mega__foo_example_org___test1__service1": "http://1.2.3.4:8080",
+			"kube_namespace1__mega__foo_example_org___test1__service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__mega__foo_example_org___test1__service1__lb_group": "",
 		})
 	})
 
@@ -793,17 +809,26 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__default_only______":                           "http://1.2.3.4:8080",
-			"kube_namespace2__path_rule_only__www_example_org_____service3": "http://9.0.1.2:7272",
-			"kube_namespace1__mega______":                                   "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__foo_example_org____":                          "",
-			"kube_namespace1__mega__bar_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__bar_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__bar_example_org____":                          "",
-			"kube_namespace1__ratelimit______":                              "http://1.2.3.4:8080",
-			"kube_namespace1__ratelimitAndBreaker______":                    "http://1.2.3.4:8080",
+			"kube_namespace1__default_only______":                                     "http://1.2.3.4:8080",
+			"kube_namespace1__default_only________lb_group":                           "",
+			"kube_namespace2__path_rule_only__www_example_org_____service3":           "http://9.0.1.2:7272",
+			"kube_namespace2__path_rule_only__www_example_org_____service3__lb_group": "",
+			"kube_namespace1__mega______":                                             "http://1.2.3.4:8080",
+			"kube_namespace1__mega________lb_group":                                   "",
+			"kube_namespace1__mega__foo_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__foo_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__foo_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__foo_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__foo_example_org____":                                    "",
+			"kube_namespace1__mega__bar_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__bar_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__bar_example_org____":                                    "",
+			"kube_namespace1__ratelimit______":                                        "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimit________lb_group":                              "",
+			"kube_namespace1__ratelimitAndBreaker______":                              "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimitAndBreaker________lb_group":                    "",
 		})
 	})
 
@@ -835,7 +860,7 @@ func TestIngress(t *testing.T) {
 		})
 	})
 
-	t.Run("has ingresses, loose a service", func(t *testing.T) {
+	t.Run("has ingresses, lose a service", func(t *testing.T) {
 		api.services = testServices()
 		api.ingresses.Items = testIngresses()
 		dc, err := New(Options{KubernetesURL: api.server.URL})
@@ -860,7 +885,9 @@ func TestIngress(t *testing.T) {
 			t,
 			d,
 			"kube_namespace1__mega__foo_example_org___test2__service2",
+			"kube_namespace1__mega__foo_example_org___test2__service2__lb_group",
 			"kube_namespace1__mega__bar_example_org___test2__service2",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group",
 		)
 	})
 
@@ -889,15 +916,23 @@ func TestIngress(t *testing.T) {
 			t,
 			d,
 			"kube_namespace2__path_rule_only__www_example_org_____service3",
+			"kube_namespace2__path_rule_only__www_example_org_____service3__lb_group",
 			"kube_namespace1__mega______",
+			"kube_namespace1__mega________lb_group",
 			"kube_namespace1__mega__foo_example_org___test1__service1",
+			"kube_namespace1__mega__foo_example_org___test1__service1__lb_group",
 			"kube_namespace1__mega__foo_example_org___test2__service2",
+			"kube_namespace1__mega__foo_example_org___test2__service2__lb_group",
 			"kube___catchall__foo_example_org____",
 			"kube_namespace1__mega__bar_example_org___test1__service1",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group",
 			"kube_namespace1__mega__bar_example_org___test2__service2",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group",
 			"kube___catchall__bar_example_org____",
 			"kube_namespace1__ratelimit______",
+			"kube_namespace1__ratelimit________lb_group",
 			"kube_namespace1__ratelimitAndBreaker______",
+			"kube_namespace1__ratelimitAndBreaker________lb_group",
 		)
 	})
 
@@ -926,7 +961,9 @@ func TestIngress(t *testing.T) {
 			t,
 			d,
 			"kube_namespace1__mega__bar_example_org___test1__service1",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group",
 			"kube_namespace1__mega__bar_example_org___test2__service2",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group",
 			"kube___catchall__bar_example_org____",
 		)
 	})
@@ -984,8 +1021,10 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__new1__new1_example_org_____service1": "http://1.2.3.4:8080",
-			"kube_namespace1__new2__new2_example_org_____service2": "http://5.6.7.8:8181",
+			"kube_namespace1__new1__new1_example_org_____service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org_____service1__lb_group": "",
+			"kube_namespace1__new2__new2_example_org_____service2":           "http://5.6.7.8:8181",
+			"kube_namespace1__new2__new2_example_org_____service2__lb_group": "",
 		})
 	})
 
@@ -1045,16 +1084,20 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__new1__new1_example_org_____service1":          "http://1.2.3.4:8080",
-			"kube_namespace1__new2__new2_example_org_____service2":          "http://5.6.7.8:8181",
-			"kube_namespace2__path_rule_only__www_example_org_____service3": "http://9.0.1.2:9999",
+			"kube_namespace1__new1__new1_example_org_____service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org_____service1__lb_group": "",
+			"kube_namespace1__new2__new2_example_org_____service2":           "http://5.6.7.8:8181",
+			"kube_namespace1__new2__new2_example_org_____service2__lb_group": "",
+			"kube_namespace2__path_rule_only__www_example_org_____service3":  "http://9.0.1.2:9999",
 		})
 
 		checkIDs(
 			t,
 			d,
 			"kube_namespace1__mega__bar_example_org___test1__service1",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group",
 			"kube_namespace1__mega__bar_example_org___test2__service2",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group",
 			"kube___catchall__bar_example_org____",
 		)
 	})
@@ -1112,7 +1155,8 @@ func TestIngress(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__new1__new1_example_org_____service1": "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org_____service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org_____service1__lb_group": "",
 		})
 	})
 }
@@ -1174,9 +1218,11 @@ func TestConvertPathRule(t *testing.T) {
 		}
 
 		checkRoutes(t, r, map[string]string{
-			"kube_namespace1__new1__new1_example_org___test1__service1": "http://1.2.3.4:8080",
-			"kube___catchall__new1_example_org____":                     "",
-			"kube_namespace1__new1__new1_example_org___test2__service1": "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org___test1__service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org___test1__service1__lb_group": "",
+			"kube___catchall__new1_example_org____":                               "",
+			"kube_namespace1__new1__new1_example_org___test2__service1":           "http://1.2.3.4:8080",
+			"kube_namespace1__new1__new1_example_org___test2__service1__lb_group": "",
 		})
 	})
 }
@@ -1309,7 +1355,7 @@ func TestHealthcheckInitial(t *testing.T) {
 		}
 	})
 
-	t.Run("use healthceck, empty", func(t *testing.T) {
+	t.Run("use healthcheck, empty", func(t *testing.T) {
 		api.ingresses.Items = nil
 		dc, err := New(Options{
 			KubernetesURL:      api.server.URL,
@@ -1513,18 +1559,27 @@ func TestHealthcheckReload(t *testing.T) {
 
 		checkHealthcheck(t, r, true, true, false)
 		checkRoutes(t, r, map[string]string{
-			healthcheckRouteID:                                              "",
-			"kube_namespace1__default_only______":                           "http://1.2.3.4:8080",
-			"kube_namespace2__path_rule_only__www_example_org_____service3": "http://9.0.1.2:7272",
-			"kube_namespace1__mega______":                                   "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__foo_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__foo_example_org____":                          "",
-			"kube_namespace1__mega__bar_example_org___test1__service1":      "http://1.2.3.4:8080",
-			"kube_namespace1__mega__bar_example_org___test2__service2":      "http://5.6.7.8:8181",
-			"kube___catchall__bar_example_org____":                          "",
-			"kube_namespace1__ratelimit______":                              "http://1.2.3.4:8080",
-			"kube_namespace1__ratelimitAndBreaker______":                    "http://1.2.3.4:8080",
+			healthcheckRouteID:                                                        "",
+			"kube_namespace1__default_only______":                                     "http://1.2.3.4:8080",
+			"kube_namespace1__default_only________lb_group":                           "",
+			"kube_namespace2__path_rule_only__www_example_org_____service3":           "http://9.0.1.2:7272",
+			"kube_namespace2__path_rule_only__www_example_org_____service3__lb_group": "",
+			"kube_namespace1__mega______":                                             "http://1.2.3.4:8080",
+			"kube_namespace1__mega________lb_group":                                   "",
+			"kube_namespace1__mega__foo_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__foo_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__foo_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__foo_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__foo_example_org____":                                    "",
+			"kube_namespace1__mega__bar_example_org___test1__service1":                "http://1.2.3.4:8080",
+			"kube_namespace1__mega__bar_example_org___test1__service1__lb_group":      "",
+			"kube_namespace1__mega__bar_example_org___test2__service2":                "http://5.6.7.8:8181",
+			"kube_namespace1__mega__bar_example_org___test2__service2__lb_group":      "",
+			"kube___catchall__bar_example_org____":                                    "",
+			"kube_namespace1__ratelimit______":                                        "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimit________lb_group":                              "",
+			"kube_namespace1__ratelimitAndBreaker______":                              "http://1.2.3.4:8080",
+			"kube_namespace1__ratelimitAndBreaker________lb_group":                    "",
 		})
 	})
 }
